Redux

+++ ClientApp/src/services/RestUtilities.js
@@ -1,15 +1,11 @@
-﻿import { AuthService } from '../App';
-
+﻿
 export default  class  RestUtilities {
 
     constructor() {
-        this.AuthService = AuthService;
     }
-
     get(url) {
         console.log("GET", url)
         return this.request('GET', url);
-
     }
     post(url, data) {
         console.log("POST", url, data)
@@ -28,9 +24,6 @@ export default  class  RestUtilities {
         let isBadRequest,
             body = data,
             headers = new Headers();
-            // stok = AuthService.getToken();
-            //  headers.set('Authorization', `Bearer ${stok}`);
-
         headers.set("Accept", "application/json");
 
         if (data) {
@@ -50,12 +43,15 @@ export default  class  RestUtilities {
             credentials: 'include'
         }).then((response) => {
             //  console.log(response)
+            if (!response.ok) {
+                throw Error(response.statusText);
+            }
             if (response.status === 401) {
                 this.AuthService.removeToken();
                 window.location.replace(`/?expired=1`);
             }
             isBadRequest = response.status === 400;
-            
+          
             let responseContentType = response.headers.get("content-type")
             if (responseContentType && responseContentType.indexOf("application/json") !== 1) {
                 return response.json();
